+++ Localization/zhCN.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2022-07-15 22-48-22 ]] --
+-- [[ Exported at 2022-07-17 15-30-14 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --