+++ CHANGELOG.md
@@ -20,6 +20,7 @@ All notable changes to this project will be documented in this file.
 - Achievement is not selected again when leaving tab and going back if it was selected before
 - Achievement objective and meta texts should now be white instead of the hard to read black
 - When in combat and using ElvUI (skinned), no more errors should appear (but skinning might not happen)
+- Clicking an event reminder that has no achievements visible due to the filters will show the empty category instead of doing nothing
 
 ## 9.2.5.36.3 - 2022-07-13
 ### Fixed